[feature] add oracle catalog for save mode
